+++ src/include/attribute.rs
@@ -1,7 +1,7 @@
 /*
     src/include/attribute.rs
 
-    Copyright (c) 2019 Stephen Whittle  All rights reserved.
+    Copyright (c) 2019, 2020 Stephen Whittle  All rights reserved.
 
     Permission is hereby granted, free of charge, to any person obtaining a copy
     of this software and associated documentation files (the "Software"),
@@ -70,5 +70,11 @@ impl Into<attr_t> for Attribute {
             Attribute::Vertical           => A_VERTICAL,
             Attribute::Italic             => A_ITALIC
         }
+    }
+}
+
+impl Default for Attribute {
+    fn default() -> Self {
+        Attribute::Normal
     }
 }