+++ src/include/attributescolorpair.rs
@@ -1,8 +1,7 @@
-
 /*
     src/include/attributescolorpair.rs
 
-    Copyright (c) 2019 Stephen Whittle  All rights reserved.
+    Copyright (c) 2019, 2020 Stephen Whittle  All rights reserved.
 
     Permission is hereby granted, free of charge, to any person obtaining a copy
     of this software and associated documentation files (the "Software"),
@@ -21,7 +20,7 @@
     IN THE SOFTWARE.
 */
 
-#[derive(Copy, Clone, Debug, PartialEq, Eq, Hash)]
+#[derive(Copy, Clone, Debug, Default, PartialEq, Eq, Hash)]
 pub struct AttributesColorPair {
     attributes: Attributes,
     color_pair: ColorPair