+++ src/ncursescolortype.rs
@@ -1,7 +1,7 @@
 /*
     src/ncursescolortype/funcs.rs
 
-    Copyright (c) 2019 Stephen Whittle  All rights reserved.
+    Copyright (c) 2019, 2020 Stephen Whittle  All rights reserved.
 
     Permission is hereby granted, free of charge, to any person obtaining a copy
     of this software and associated documentation files (the "Software"),
@@ -81,8 +81,7 @@ pub(in crate) fn set_ncurses_colortype(colortype: NCursesColorType) {
         } else {
             EXTENDED_COLORS_ALREADY_SET.store(true, Ordering::SeqCst);
 
-            EXTENDED_COLORS.store(
-                match colortype {
+            EXTENDED_COLORS.store(match colortype {
                     NCursesColorType::Normal   => false,
                     NCursesColorType::Extended => true
                 },